renamed some tests and developed some code path coverage information in the test worksheet.
i need to finish the logic for redistributeKeys.  there are no tests at the node level.

i think that removeChild is not quite right and that merge is not quite right either as a result.  there are notes on this issue in the code.  I am not sure if there is correctness at stake as opposed to only elegance.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@72 8f7bc0f5-282e-0410-95e3-8d296e9bb460