Wrote a stress test for the IndexSegmentBuilder.
Added a reverse term index (termId:term) to the RDF test class.  This dropped the throughput by 50k triples/sec down to 120k triples per second.

I added a “reverse index” to map term identifiers (long integers assigned to RDF values) to terms (the RDF values).  This brings down the sustained insert rate to ~120,000 per second.  That is a cost of about 50,000 statements per second.

In order to support query we need to add either link set style metadata or indices other than SPO.  I am inclined to try both.  Adding more indices is the most straight-forward.  I am not quite sure how I would want to handle the link set style metadata.  This introduces a dependency between the term (RDF Value) and the Statement.  You need to resolve the terms, add the new statement set its metadata while updating the statement metadata for the statement that was the tail of that link set, and then update the #of items in the link set and its tail on the term.  This is easy in the GOM since all of that is in the object cache.  At the B+-Tree level you do not have that sense of locality.  I am thinking that a feature like a “finger” that sticks on a point in the index would help to restore that or a batch oriented approach to doing this that takes advantage of the caching of nodes and leaves in the index….

There are a lot of things that I can do to bring the performance back up: better write buffering, sorting keys and doing batch inserts, changing the SPO keys to be long[] (vs Long[]) and using a stride of 3 (for a triple store) or 4 (for a quad store).  Those would be my top items.  I am tempted to try multiple indices now and see what the throughput looks like since that is one reasonable architecture.  I am hoping that it will not bring us down below 50k per second.  And then there is inference.  The other place to pick up more performance will be on parallel loads with multiple machines.  I’m not too worried – I think that we are definitely in the ballgame.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@89 8f7bc0f5-282e-0410-95e3-8d296e9bb460