Modified RuleRdfs11 to use a binary search to locate possible matches during the self-join.  This gives a big speedup on the nciOncology dataset.
There is more work to do here.  I think that RuleRdfs11 should fix point all by itself in order to compute the transative closure of the subClassOf relationship as quickly as possible (in as few rounds).  This the join is handled in the abstract rule shared by the subPropertyOf rule, this will also provide a fast fixed point for that relation.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@181 8f7bc0f5-282e-0410-95e3-8d296e9bb460