Refactored the KeyBuilder into a base class without Unicode support (light weight) and a UnicodeKeyBuilder (heavier weight and not always required).
Modified the BigdataClient to cache the metadata index locally and moved to a static partitioning scheme for scale-out indices as a first step.  I am still working through the client API for transparent access to the scale-out index partitions and then I will update the rdf application to use the client API.  I have not yet done the asynchronous overflow handling or the quad store migration.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@209 8f7bc0f5-282e-0410-95e3-8d296e9bb460