This commit is a mostly running refactoring of the triple store to use unisolated indices and to operate against a local database, a temporary database (for inference), or a bigdata federation.
There is some issue with containsStatement(...) that shows up during inference.

The test suite tends to exhaust resources such that some tests fail with out of memory errors that would succeed if run directly.

The performance appears to be down from ~20k triples per second to ~13k triples per second.  I need to look into this further.  The main change was to how bulk loading of terms into the lexicon is performed.  (Also, the bulk loading of the statement indices has not been realized yet).  One possible explaination is that there is more overhead with the UnisolatedBTree, but I suspect that we can improve on this performance.  In any case, this design should now support concurrent writers (not tested).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@218 8f7bc0f5-282e-0410-95e3-8d296e9bb460