Defined an embedded variant of the partitioned federation and ran some tests comparing native use of the journal (TripleStore), and the ScaleOutTripleStore using the embedded and distributed federations.
Performance of the embedded federation is about 60% of the native use of the journal (10,000 tps) while performance of the distributed federation is about 40% of the native triple store (so 20% of the cost is network IO, which could be addressed by compression during serialization of key/value buffers).

Not forcing on commit raises the performance of the embedded federation to 72% of the native journal.  Group commit will allow us to recover performance here for _concurrent_ clients, but it will not provide a gain for a single client.

Overall, the ClientIndexView does more work than the BTree and I have not yet isolated the costs therein.

Also, the ScaleOutTripleStore has special remote procedures to handle consistent inserts into the terms, ids, and statement indices and I have not yet tried to trim down the costs in those procedures, e.g., extra data allocation or copying, etc.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@226 8f7bc0f5-282e-0410-95e3-8d296e9bb460