Major refactoring to allow addressing of more than int32 bytes in a single store.
The packing of addresses in the btree nodes is broken.

This adds some Apache License (v2) code (SystemUtil) that was abandoned by its project (Excalibur/fortress).  SystemUtil can report the #of cpus on the host platform, which makes it easier to parameterize some things.

There is a map/reduce implementation running on a local federation - it needs performance tuning.  This is a good place to work on group commit.  Note that group commit will also require modifying indices to not participate by default in a commit by the journal.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@236 8f7bc0f5-282e-0410-95e3-8d296e9bb460