Added support for creating temporary files to the embedded federation.
Fixed a bug in map/reduce where the hash partition was being computed on the total key rather than just the application key.  The result was that the additional information (map task identifier and tuple counter) in the key was causing tuples for the same application key to be split into different reduce partitions (which is wrong).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@237 8f7bc0f5-282e-0410-95e3-8d296e9bb460