Ractored out an ITransactionManager, IResourceManager, and an IConcurrencyManager.  Wrote an IResourceManager that handles journal overflow.
Removed ReadComitted as a kind of transaction and replaced its use throughout with a constant, ITx#READ_COMMITTED.  This constant has the same semantics.  Removed the readOnly parameter from AbstractTask and derived that information from the timestamp (less than zero is always read only) and whether or not the timestamp is a read-only transaction.

(Re-commit)


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@509 8f7bc0f5-282e-0410-95e3-8d296e9bb460