Modifed to use the overflow handler as appropriate when copying indices during overflow processing from the old journal to the new journal.  This let's indices that register an overflow handler copy forward any raw records that they reference (aka blob references).
Added a limit on the #of non-zero indices that can be copied during synchronous overflow processing.  This is to catch conditions where a large #of indices are registered on a journal with only a few records each.  We want to avoid copying too many indices to the new journal since it will increase the latency of the overflow event.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@525 8f7bc0f5-282e-0410-95e3-8d296e9bb460