Fixed a ClassCastException during BuildIndexSegmentTask$AtomicUpdate. This was being thrown due to a pre-condition assumptions.  The assumption was that the source view would include at least two sources - the B+Tree on the new journal and at least a B+Tree on the old journal as well.  I am going to assume for now that there are valid pre-conditions where those assumptions do not hold and rely on a ground truth correctness suite to validate the overflow processing.
Fixed KeyAferPartition exception.  This was arising for some byte[][]
keys when the keys contained 2 or more entries which compared as EQ to
the rightSeparator key for the index partition.  The fix was to
explicitly back up until we are sure that the toIndex for the split
falls on the first key that is equal to the rightSeparator and not on
the 2nd or 3rd key.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@544 8f7bc0f5-282e-0410-95e3-8d296e9bb460