Added counter set reporting for the live journal.
Enforcing read-only semantics.

closeForWrites()

      - Does NOT close the journal to avoid problems for concurrent readers

      - Discard the write cache in order to reduce the memory footprint.

Fixed a write cache synchronization problem. The bug surfaced the change to the group commit.  Since it now checkpoints indices and allows writers to continue during commits there are now concurrent writers when we call force.  The problem was that DiskOnlyStrategy#force() was not synchronized on the buffer strategy.  AbstractMRMWTestCase was modified to demonstrate this problem by sometimes invoking IRawStore#force().


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@577 8f7bc0f5-282e-0410-95e3-8d296e9bb460