Statements about statements are now retracted if the statements that they are about are retracted.
Some of the test suites were using constants (1L, 2L, 3L) that caused problems since the term identifiers are now bit flagged and some of the constants were being interpreted as statement identifiers (3L).

I also cleaned up a lot of the stuff to handle removal of statement and justifications, both of which should be significantly faster now.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@590 8f7bc0f5-282e-0410-95e3-8d296e9bb460