working on infinite loop problem in TM
changes dumpStore() to return a String and modfied callers to write out that string on the console or logger.

found bug in TMStatementBuffer where the ungrounded and downgrade buffers were conditionally flushed based on the logging level.

modified logging in TMStatementBuffer that used SPO.toString(db) to be conditional on the logging level to avoid database looks during unlogged execution.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@591 8f7bc0f5-282e-0410-95e3-8d296e9bb460