fixed problem with nano http where it would fail to report a error in the serve() method.
fixed problem with history where it was throwing an assertion if there were more samples that slots with values - this is possible now that a slot may have more than one sample.

made the date format a member field for sar and pidstat - i think that there were thread-concurrency problems since they shared one instance for the whole jvm.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@669 8f7bc0f5-282e-0410-95e3-8d296e9bb460