fixed jini federation initialization - the load balancer and timestamp service discoverty were not being properly initialized.  also the data service was not notify()ing the LBS on startup.
fixed problem with sar and pidstat having to do with field overflow.  The data lines are now split based on whitespace.

change the concurrent data loader default to autoFlush=true.

modified the store manager to not assert that the live journal is read-only since that assertion is temporarily violated during synchronous overflow and made the liveJournal an atomic reference.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@696 8f7bc0f5-282e-0410-95e3-8d296e9bb460