Bug fix for corrupt index segment build - the disk only strategy was failing to flush the write cache before initiating the transfer.
Identified limit of 1.5-2G for file channel transfers under windows - known JVM bug.

Refactor of TemporaryRawStore and DiskOnlyStrategy to provide a new temporary store facility.  There is now a "Temporary" journal mode (supports full concurrency).  The new TemporaryRawStore directly uses a DiskOnlyStrategy so it inherits the MRMW characteristics.  Instead of an overflow from a Transient buffer to a DiskOnlyStrategy, the DiskOnlyStrategy is now able to defer the creation of the backing file until the first write to the disk. It also understands other semantics for temporary stores, including not writing root blocks and not forcing writes to disk.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@718 8f7bc0f5-282e-0410-95e3-8d296e9bb460