Fixed bug in FileChannelUtility#readAll() - it now detects reading at/beyond the end of the file.
Added prior/next double-linked leaves to the IndexSegmentBuilder.

Added ILeafIterator and an implementation for the IndexSegment that uses the prior/next leaf references -- this still needs to be integrated into the AbstractBTree in order to get utilized and that means writing an ILeafIterator for AbstractBTree that can handle prior/next leaf visitation patterns.

Defined a new interface IUpdateStore and put the allocate(int nbytes) and update(addr,off,ByteBuffer) methods on that and declared that interface for TemporaryRawStore and DiskOnlyStrategy.

Renamed the index segment dump utility.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@719 8f7bc0f5-282e-0410-95e3-8d296e9bb460