Refactored DataService to define getFederation() and refactored the access methods for the timestamp service, metadata service, etc. as calls to the federation.
I had to chance the DataService startup as well.  You now issue a start() after the ctor in order to begin operations.  (There was a problem was the ctor of a subclass not being fully initialized during method calls originating in the base class ctor that were at least in part dependent on the uninitialized state of the concrete impl).

This change simplies things a bit.  The jini services are not up and running yet with this change as it impacts their jini config files.  Those files will have to be modifed to account for the ClientDescription vs ServiceDescription vs AdventDescription.

I am still working on executing the rules inside of the data service.  It runs the query fine, but there is a problem when running the insert.  i suspect this is related to the locks.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@762 8f7bc0f5-282e-0410-95e3-8d296e9bb460