refactor introduces an IKeyBuilder to the IndexMetadata and a factory for the same.
added atomic row delete for the sparse row store.

cleaned up IRelationLocator and deleted IRelationFactory (redundent).

moved the RDF relation and rules stuff into the RDF DB module.

fixed a problem with asynchronous iterators for access paths and rule evaluation (the future was not being monitored by the iterator).

made the global row store concept universal.

the global row store is now used to store the relation declarations, including the runtime class that we need to instantiate a relation.

there are still changes to the jini package that have not been checked in dealing with the refactor of the data service as a full on federation client.

simplified the resource manager to use getFederation() rather than several other methods.

added a getLastCommitTime() to the federation.  this will be used to support read-consistent operations, primarily read-consistent iterators.

added a NamedLock class to support very simple named resource queue models.  this needs to get applied to the StoreManager as well to reduce latency when opening a store (it is synchronized on the method rather than on the resource being opened).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@769 8f7bc0f5-282e-0410-95e3-8d296e9bb460