Major changes for the rdf db to support efficient joins both in the data service and distributed across the federation and to remove the memory limit from joins in the LocalTripleStore
THIS COMMIT BREAKS FUNCTIONALITY

In particular, closure and truth maintenance have obvious bugs that are linked to the read-behind from the focusStore.  Read behind was introduced in order to have greater concurrency.  The problem just needs a little more working through.  Mainly, it has to do with how we locate and materialize views of the relations of interest and the fact that READ_COMMITTED semantics for a BTree are more like read-historical from the last commit point, whereas they are more "read-committed" for the data service layer.  Also, there is a problem with caching of read-committed relation views, which is really the same problem.

There is also what appears to be a pre-existing bug with statement identifiers.  When enabled, a document with a lot of bnodes can overflow the buffer and cause an incremental write where some bnodes have not yet been resolved to sids.  When disabled, it appears that parsing documents with bnodes leads to an unbound (0L) term identifier in the object position -- clearly the statement identifier semantics are interferring with the normal RDF XML handling.

Performance for the text indexer seems slow, at least on some data sets.  Performance with it turned off seems ok.

Much of the test suites run fine.  There are a few errors that appear to be linked to resolution of relations and one in the backward chainer that might be that or something else altogether.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@783 8f7bc0f5-282e-0410-95e3-8d296e9bb460