(broken pipe, recommit)
working through filtered iterator stacks for tuples based on MartynC's striterator package.  many of the iterator related issues have now been fixed, but there is still some more work to be done on the partitioned and chunked variants for reverse traversal and some of the filters for the bigdata repository, etc.

refactored the key encoding and decoding, the index procedures, and the tuple filters into their own packages.

this commit seems to have re-introduced some problems in the backchainers (may actually be related to the recent change to also scan the POS index for x type resource) and the RDF DB closure.  i'll look into those once I finish the iterator changes.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@839 8f7bc0f5-282e-0410-95e3-8d296e9bb460