Made the InferenceEngine class flyweight, refactored the Axioms model and how it is declared, and refactored the Vocabulary model and how it is declared.
AbstractTripleStore#create() is now responsible for creating the
Axioms instance and the Vocabulary instance for the triple store and
saving them as properties in global row store.  Those properties are
fetched from the global row store on demand.

Loading axiom model into database when it created required updates to
a bunch of unit tests.

Dropped [rdfsOnly] option in favor of the explicit selection of the
Axioms class.

Dropped Fast and Full options in favor of the explicit selection of
the BaseClosure class.  Changed Fast and Full to be classes
(FastClosure and FullClosure) that generate programs.  This is
extensible, even if the implementations need to be aware of a bunch of
options related to the inference engine and what is to be forward
chained vs backward chained.

Modified the DoNotAddFilter to (de-)serialize the Axioms with the
filter.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@889 8f7bc0f5-282e-0410-95e3-8d296e9bb460