Made sure that everyone is using IIndexManager#getTempStore().  This automatically rolls out a new TemporaryStore once the old one begins to get too large.
Modified the backchainers to NOT accept a TemporaryStore parameter.  They all use AbstractTripleStore.getIndexManager().getTempStore() to obtain the (current) temporary store on demand.

Refactored BigdataSolutionResolverator.  It now is a general purpose utility to resolve bigdata bindingsets containing term identifiers to the referenced BigdataValue objects from the lexicon.  There is also a glue iterator class to align our iterators with Sesame 2 closeable iterations.

Fixed provenance query using native rule execution.  The SPOPredicate did not have a context position so the SID (when present) was not being copied into the binding set.

AbstractTripleStore#match() now resolves term identifiers in the binding sets to the BigdataValues.  However, the unit tests are still failing because the native rule execution is causing the "lit" variable to drop out when we specialize the matchRule for a specific literal.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@898 8f7bc0f5-282e-0410-95e3-8d296e9bb460