Added owl:inverseOf and owl:TransitiveProperty and additional axioms to support those rules.  Currently, these rules are not enabled by default.  When they are enabled, we now get LUBM Query 11 right, but that's all the new answers that are available.  Those additional answers are all due to TransitiveProperty - inverseOf is not being used.  We would need entailments from owl:Restriction to do better than this.
Added parallel execution of subqueries for nested joins.  At
present, parallel subquery is restricted to the first join dimension
and it issues subqueries in parallel for each element in each chunk of
that join dimension.  Subsequent join dimensions run in the thread of
the caller.

Added logic to re-order the chunk from the current join
dimension into the order for the next join dimension in order to
improve the use of ordered reads in subqueries.

Modified BlockingBuffer to use poll() vs sleep() for less latency.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@903 8f7bc0f5-282e-0410-95e3-8d296e9bb460