found a nasty penalty for the federation use cases in DefaultResourceLocator.  it was failing to cache the read-committed resources views because of an issue with caching them for a resource using com.bigdata.btree.BTree.  a 3x win in performance!
normalized options for turning on and off platform statistics and queue statistics collection and the httpd service for reporting the same into IBigdataClient.Options.  The LoadBalancer now uses the same properties to configure its httpd service, so you MUST override the port for the load balancer if you want it to run its httpd service on a known port.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@917 8f7bc0f5-282e-0410-95e3-8d296e9bb460