A variety of performance enhancements.  LTS, LDS, and EDS now have nearly the same data load and closure rate.  JDS is about twice as slow.  LTS and LDS query are both fast.  EDS query is slower.  JDS query is painful - there is some problem with the BlockingBuffer and the async proxied iterator that I have not yet worked out.
Changes include: Range counts were not cached.  Index metadata was not cached (for LDS).  The use of a cached thread pool was causing the work queue to be a synchronous queue and that broke the logic to delay tasks when the work queue was nearing its capacity (the synchronous queue has a zero capacity).  The Journal's ResourceMetadata was not being cached and that was leading to significant overhead so I dropped the size() method and now IResourceMetadata is invariant once the store is created.  Plus others.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@920 8f7bc0f5-282e-0410-95e3-8d296e9bb460