A variety of possible performance tweaks.  These changes include a new lgpl-utils JAR (1.1-b3).
Done. Modified DefaultResourceLocator to test the cache without a
named lock.  On a miss it acquires the named lock and then verifies
that there is still no entry in the cache once it holds the lock.  If
the resource has not been concurrently located, then it is located and
placed into the cache and the lock is released.  This addresses a hot
spot where acquiring the named lock was too expensive.

Done. Modified Node#getChild(int) do not declare itself as
[synchronized] and instead use a double-checked locking pattern.
Node#childRefs was marked as [volatile] in support of that change.
This addresses an what appears to be a hot spot in Node#getChild(int)
where the self-time was higher than could be accounted for by the code
in the method, leading me to believe that the synchronized keyword was
causing contention for access to the method for Node instances.

Done. Modified SPORelation to check the access path cache first, then
synchronize on a lock object iff there was a cache miss and do the
access path factory behavior and update the cache while synchronized
on that lock object.

Done. Each child in the B+Tree has its own weak reference to the
parent.  If we put a weak reference on the parent to itself then the
child could reuse the same weak reference object.  This would cut down
on the #of weak reference instances drammatically and that might help
the garbage collector.

Done. Modified InputBitStream and OutputBitStream to have ctor
variants that avoid a reflection test on the underlying stream.  This
was causing approximately 5000 MethodNotFoundException instances to be
thrown per second.  Introduced a new lgpl-utils jar for this change.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1007 8f7bc0f5-282e-0410-95e3-8d296e9bb460