Symptom. Asynchronous shutdown of the executor service while a query was executing would cause an ERROR to be logged and an RejectedExecutionException to be thrown.
Cause: When normal shutdown of the service is requested it is common
that the main thread will be in a state in which it attempts to
schedule more task(s). This results in a RejectedExecutionException.

Resolution: We treat this just like an interrupt since the join can
not progress due to the shutdown of the executor service.  The
RejectedExecutionException is masqueraded as an InterruptedException
and the NestedSubqueryWithJoinThreads task already knows how to handle
the InterruptedException.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1009 8f7bc0f5-282e-0410-95e3-8d296e9bb460