Bloom filters and also turned on the read retention queue for BTree and hacked the SPORelation to disable the access path cache for the unisolated view (just testing something out there w/ mike)
Done. Modified AbstractBTree, BTree, IndexSegment and friends to
permit a bloom filter to be maintained on a BTree as well as on an
IndexSegment.  This is valid since keys removed from the BTree will
result in false positives that will be disproved when tested against
the data in the index.  This is being done in order to explore the
utility of a bloom filter on the SPO index as a means to reduce costs
for complex joins involving lots of point tests (LUBM Q9 has three
joins that are point tests, as does LUBM Q2).

Done. Experiment with the use of a bloom filter for the SPO index (the
one that is used for a point access test).  Can this reduce costs for
complex joins such as LUBM Q2 and LUBM Q9?  Yes, but there are scaling
limits so this works best with an index view (LDS) or key-range
partitioned index (EDS/JDS).  For LTS, the bloom filter will be
shutoff automatically once the index size would cause the bloom filter
error rate to be unacceptable.

Done. Modified the BTree unit tests to add tests in which the bloom
filter was enabled.

Done. Need flag for whether the bloom filter is dirty and must test
that flag in writeCheckpoint() and write out the bloomFilter and
update its address if it was dirty, then reset the flag.  I also need
to make sure that the bloomFilter is getting persisted and that its
address is being updated, and that the dirty state for the bloom
filter is reset when we create a new bloom filter.

Done. Finish changes in AbstractAccessPath (needs API for counter
updates).

Done (for BTree). Consider lazy materialization of the bloom filter to
reduce cost for code paths that do not touch it.  This way we can use
a (somewhat) larger bloom filter and not pay a performance penalty for
queries that don't use it (mainly this is gaming the lubm test harness
since it re-opens the repository for each set of query presentations,
but it could have value elsewhere as well).

Done. Track false positive rates for the bloom filter.

Done. If the bloom filter's false positive rate is too high for an
index then stop testing the bloom filter for that index object.  If
the index is a BTree, then drop the bloom filter so that we no longer
have to read it from the disk, update it, or write its updated state
on the store.  However, do not disable the bloom filter option for a
scale-out index since the performance of the bloom filter will be good
again once we overflow.

Done. Fine tune the bloom filter serialization.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1039 8f7bc0f5-282e-0410-95e3-8d296e9bb460