Commit notes:
Note: Closure is not still not correct for LTS/LDS/EDS on pipeline
joins.  In all cases the pipeline join overgenerates by the same #of
inferences. I will debug for LTS first since this is easier.

Done. I was failing to set the predicateId on the SPOPredicate and
then I was failing to use the new SPOPredicate with the predicateId.
This meant that the pipeline was always using the scale-out indices.

Done. RDFJoinNexus#getTailAccessPath() was not calling init() on the
SPOAccessPath.

Done. Was not tracking the mutationCount for the pipeline join in
JoinTask#flushAndCloseBuffersAndAwaitSinks()

Done. Run LTS closure as UNISOLATED.  This should work for both nested
and pipeline joins, but we have to make JoinTask a simple callable
first and make sure that the MutationTask sets up all necessary lock
requests for the joins (it should since it already does this for
nested subquery).

Done. Run LDS/EDS/JDS closure as read-consistent and advance the
readTimestamp before each mutation task to the lastCommitTime.

Done. EDS closure is slow because we are not aggregating the solution
chunks which is resulting in a large #of sequential inserts of small
chunks on the relation.  [Modified AbstractSolutionBuffer#flush(...)
to aggregate the caller's chunks into a single perfect fit chunk.]

Done. LDS closure and query run now that JoinTask is no longer an
AbstractTask.

Done. Fix LDS query (read-only access to the row store was throwing an
UnsupportedOperationException from
AbstractJournal#getIndex(name,timestamp)).  A commitTime was being
passed in as a timestamp without using
TimestampUtility.asHistoricalRead(commitTime).

--------------------
Places of interest for JOINs.

AbstractTripleStore

	-newJoinNexus : choice of the readTimestamp.

RDFJoinNexusFactory

	- indexManager:joinNexus cache.

	- setReadTimestamp(timestamp)

RDFJoinNexus

	- runDistributedProgram(IStep)

ProgramTask

   Advances the readTimestamp before running each mutation step.

	executeMutation()

Future<RuleStats> AbstractStepTask:submitToConcurrencyManager()

   Chooses the timestamp for the AbstractTask.

DefaultRuleTaskFactory

   Responsible for using either the JoinMasterTask or the
   NestedSubqueryWithJoinThreadsTask.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1054 8f7bc0f5-282e-0410-95e3-8d296e9bb460