Pipeline closure and nested subquery closure now agree.  They are also very close to one another in performance except for the EDS condition where the pipeline closure is 2.5x faster than the nested subquery closure.  Hopefully that bodes well for the performance of the pipeline closure for scale-out.
Done. [The DoNotAddFilter was not being applied and so literals were
making it into the subject position.] Closure is not still not correct
for LTS/LDS/EDS on pipeline joins. In all cases the pipeline join
overgenerates by the same #of inferences.  Debug for LTS first since
this is easier.

Done. Refactored the model comparison stuff into a utility class.

Done. Fixed bug in AbstractStatementBuffer where an if() test was not correctly parenthesized.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1055 8f7bc0f5-282e-0410-95e3-8d296e9bb460