Refactored the pipeline join into its own package.
Obtained baseline JDS join performance for LUBM U1, U5, U10, and
U50 on a single machine (laptop).  LTS is only 1.3x faster for closure
and 3x faster for query.  Those estimates are on a resource
constrained laptop.  JDS should do much better on a cluster of server
class machines.

Modified the IndexSegmentStore to transparently re-open the
FileChannel if it is closed by an interrupt during an operation on
that channel.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1059 8f7bc0f5-282e-0410-95e3-8d296e9bb460