Added a transient mode for the BTree class and a test suite.  See BTree#createTransient(...).
Possibly fixed the problems with resource manager overflow which were on the one hand leading to dropped index partitions during purge (it was not running forwards through all known commit points in order to build up the actual set of dependencies required for views as of the commit time to be preserved) and on the other hand was not able to produce a truely compact footprint.  However, even if the new solution is correct (remains to be tested at scale) it may take too long when there are a large #of commit points to be scanned, in which case we might have to maintain an intermediate data structure specifically for this purpose.  I am logging the elapsed time for this computation and we will see how it looks for larger data sets.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1135 8f7bc0f5-282e-0410-95e3-8d296e9bb460