mostly fixed the historical timestamp issue and I am now working on extending read-only tx support throughout the system.
added logic to the data service to periodically clear stale references for indices and stores, but this is not really complete yet as it does not do the locator cache, the caches on the live or historical journals, etc.

changed the namespace for BFS indices to conform with guidelines.

added read-locks to the IndexManager.

Journal#newTx(long) is broken, but that is the next thing on my list.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1138 8f7bc0f5-282e-0410-95e3-8d296e9bb460