Done. Generate and visualize a concise summary of how the disk space in the federation is allocated (a) on each data service on each host; and (b) by index.  Also show the #of sources in each index partition. The format is a tab-delimited.  It can be sorted in any number of ways to summarize the information as desired.
Done. Bug fix for FusedView.  It was failing to substitute the actual
left and right separator keys when none were specified for rangeCount
and rangeIterator.  This is a problem because an IndexSegment is
reused after a split by both of the new index partitions.  Failing to
constrain the key range when no constraint was specified was resulting
in counting / reading too many tuples since tuples outside of the
index partition boundaries were being included.

Done. Added separate counters for journal bytes under management and segment bytes under management for the store manager


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1281 8f7bc0f5-282e-0410-95e3-8d296e9bb460