Done. Modified to track BTreeCounters for each index partition using a single BTreeCounters instance per index partition.  The instance is set on the AbstractBTree by IndexManager#getIndexOnStore().  This makes it canonical and gets rid of the memory leak in ViewCounters.
Done. Modified to consider only the write time when computing the
Score for the previous journal during overflow processing.  This let's
us use canonical counters but focus MOVEs on index partitions which
are hot for writes.

Done. Modified OverflowManager and DiskOnlyStrategy to add in the
counters from the old journal during overflow so we now track the
cumulative counters for the live journals over time.

Done. Fixed problem where the live journal counters were not updating
after an overflow (it was trying to reattach them to the local
CounterSet object, but the CounterSet had attached as a child of the
service root so we needed to be operating on the service root
instead).

(*) Done [Needed to detach and then attach rather than reattach as
    that does not remove the existing counters.]  The btree counters
    are not being removed from the LBS (they won't) or the data
    service httpd (they should).

(*) Done [AbstractTask needs to set the BTreeCounters when it is first
    created and when it is reloaded]. The structure counters for the
    BTree seem to be always zero (leafSplits, tailSplits).  The other
    counters may be reporting low.  It is as if the mutable BTree's
    counters are not being set to the same object as the read-only
    BTree's counters.

(*) Done [AbstractTask needs to set the BTreeCounters when it is first
    created and when it is reloaded]. Sometimes see no performance
    counters on hand for an index which exists, but not always.  I bet
    that this is related to not seeing any performance counters for
    the mutable BTree.  If there has never been a read-only view of
    the BTree then the performance counters are missing.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1328 8f7bc0f5-282e-0410-95e3-8d296e9bb460