Finally fixed the term identifier encode/decode logic.  It is now parameterized so that we can use N bits from the local counter and reverse them and place them into the high N bits of the encoded term identifier.
Note: This has BROKEN binary compatibility with existing stores since
the long term identifiers are no longer guarenteed to be non-negative
and therefore are no longer being packed.  There is a CVS TAG for the
previous codebase state: BIGDATA_PRE_TERM2ID_VALUE_ENCODING_CHANGE.
You can export as RDF/XML and re-import the data across this codebase
change.

Next step is to try measure the effect of different parameterized
values, e.g., 0 bits (should be identical to the historical scale-out
performance, excepting only that we are not packing the term
identifiers); 6 bit (should cause writes to be distributed across the
index partitions while retaining good locality within an index
partition); and 31 bits (should cause writes to be distributed across
the index partitions but should cause the within index partition
access to degrade since the writes will also be uniformly distributed
within an index partition).

Added com.bigdata.jmx.JMXTest which demonstrates how to register
MBeans using StandardMBean, an interface, and an object which
implements that interface.  This is much simpler than using the
abstract classes for dynamic mbeans, etc.

Added a utility method that can be used to register the log4j logger
hierarchy as an MBean.  This was refactored from the apache zookeeper
project.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1420 8f7bc0f5-282e-0410-95e3-8d296e9bb460