done. AbstractBTree#writeTuple now uses both KEYS and VALS since otherwise insert() and remove() are unable to reconstruct an application object which is partly encoded within the key.
continued work on a refactor of the performance counter system.

done. Introduced "hot splits" as complementary to moves and taken  When a service is highly utilized (as identified by the LBS) but not
resource constrained.  The purpose of the "hot split" is to split hot indices and thereby increase the potential concurrency.  Moves are complementary because they are also taken when the LBS identifies the service as "highly utilized" if the service IS resource constrained.

done. Fixed references to resources in the file system which were no
longer valid when running the unit tests from the master directory for the project (these are still file references, so the tests will fail if you check out an individual module and run the tests within that module).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1441 8f7bc0f5-282e-0410-95e3-8d296e9bb460