some work on deflate integration for the NodeSerializer (not finished yet).
added a huffam serializer for btree values based on mg4j (MikeP) and integrated into the ID2TERM index.

Bug fix to SerializerUtil when ByteBuffer was backed by an array and the position (pos()) was non-zero.

Modified Node to dimension childLock[] to (nkeys+1) rather than (branchingFactor+1).  We were allocating objects for the elements of this array which were not reachable by B+Tree navigation for a read-only BTree.  There is a lot of similar fat that needs to be trimmed from Node's internal arrays when used with a read-only BTree view.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1526 8f7bc0f5-282e-0410-95e3-8d296e9bb460