modified async index write stats to report the average sink queue length rather than the average total of the sink queue lengths
modified blocking buffer to self-report the #of elements in the queue when the generic type is an array type.

modified the RDF async loader to NOT store the BigdataValue reference with the KVO for
the ID2TERM index and to NOT store the BigdataStatement reference with the KVO for the
statement indices.  This should help to reduce the retained heap.  The reference is
stored with the KVO for the TERM2ID index since we need the term identifier, which is
assigned by a side-effect on the BigdataValue for that index.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1629 8f7bc0f5-282e-0410-95e3-8d296e9bb460