Added some embedded retry logic for the async write api.  it will now retry up to 3 times per write before giving up, at which point the exception will be set on the master and the master will be closed.  this should eventually be refactored to be an extensible retry policy.  See my notes on that in the code.
Modified IndexSegmentStore#reopen() to not throw an exception if the store is already open.  IndexSegment#readNodeOrLeaf() does not have a lock when it invokes reopen() so a concurrent reopen is possible and was causing an IllegalStateException to be thrown from reopen().  reopen() now returns without throwing anything if the store is already open once the caller holds the lock.

Modified the archiveRun.sh script to archive rolled over logs as well as the current log for the events log, detail log, and error log.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1667 8f7bc0f5-282e-0410-95e3-8d296e9bb460