Added clear() to BlockingBuffer to clear the backing queue, but this is only allowed if the buffer has been closed (this is to avoid introducing possible concurrency problems).
Modified the asynchronous write API to clear the backing queue for the master and the sink if the job is cancelled.

Added documentation to the bigdata configuration file to the effect that ParallelOldGC requires jdku10 or above due to a memory leak.  This also corresponds to the U100000b/run9 configuration.

Added some documentation of examples in the bigdataprecond file.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1693 8f7bc0f5-282e-0410-95e3-8d296e9bb460