Fix to archiveRun.sh (bash syntax error).
Modified the IndexSegmentStore to report the index and file when the nodes region exceeds the available extent of the direct byte buffer pool. This makes the warning that it logs more useful.

Configuration files modified to use 2M direct byte buffers for the common direct byte buffer pool (1M buffers were insufficient for index partitions that are approaching their maximum size on disk as the nodes region is approaching 2M at that time).  In the future, there should probably be a set of direct byte buffer pools of different sizes / purposes such that we use only as many as a required of any given size and can ensure that the index segment nodes remain fully buffered regardless.

Modified the master and sink queue capacities for the 3-node/32-bit configuration and the 16-node/64-bit configuration to permit more buffering of data.

Modified the default deployment log4j properties file to turn off some log messages and rationalize its layout in the configuration file by grouping related properties together.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1715 8f7bc0f5-282e-0410-95e3-8d296e9bb460