Modified the batch performance counter extraction utility to ignore files with XML errors as a workaround since partial files can be captured if they are copied while they are being generated by the LBS.
Removed javadoc on problems with the live journal counters from the DiskOnlyStrategy since that issue is now resolved.

Modified the AsynchronousOverflowTask and the OverflowManager to define and use a configurable timeout that the service will await an exclusive write lock in order to purge old resources before giving up.  This addresses an issue where the bytes under management would grow without bound on services which were under a sustained heavy write load.  The default timeout is 60s (it was hard wired at 3s).  An error is now logged if the lock can not be acquired and purge resources does not run.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1736 8f7bc0f5-282e-0410-95e3-8d296e9bb460