Dialed down the heap size in misc processes both in the configuration files and in the scripts (bigdataup, lubmMaster.sh). This matters most for machines with a lot of RAM as the JVM will default to a much larger heap that is in fact required.
Javadoc edits on the asynchronous write api classes.

Changed the poll timeout for asynchronous write sinks from 50ms to 2000ms in the configuration file.  I think that this was stealing CPU time once there were a large #of index partitions (1000s) on which the clients were trying to write.   Just too many threads doing short polls.  While each thread was polling its own buffer, this led to scheduling contention.

Turned on large page support in the 16-node configuration file for evaluation.

Setting vm.swappiness=0 fixed swapping on long runs.  The linux kernel was trying to keep lots of RAM available for applications which might want to run, with the result that the applications trying to run were not able to use ~50% of the available RAM.

Modified the archiveRun.sh script to pre-extract the counters for post-mortem analysis since this takes 2m on a server and much more time on a workstation.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1756 8f7bc0f5-282e-0410-95e3-8d296e9bb460