Modified the asynchronous write API to no longer await the completion of a subtask whose buffer has been closed when the master needs to transfer another chunk to a sink associated with the same locator. This was being done historically to ensure that we always test the Future of the subtask in order to detect errors in those tasks and halt the master on error.  The subtasks are now placed onto a queue when their notify the master than they are done processing.  The master polls that queue in call() and in awaitAll().  If the next subtask in the queue is done (Future#isDone() is true), then the Future is tested at that time.  This prevents the master from blocking on a chunk transfer when the old subtask for that locator is closed but its Future does not become available for 15s+ because it is doing RMI and waiting on an index partition write to complete.
In a manner similar to the refactor to use a redirectQueue, this can
had broader implications in the code with respect to termination
conditions for the master.  Also, I have made most of the asynchronous
write API counters into AtomicInteger and AtomicLong fields in order
to have atomic visibility change guarantees for some of the unit tests
which otherwise could examine the state of the counter from a
different thread and potentially read the "wrong" value.

There is still one fence post in some of the unit tests. The symptom
is that the unit test fails due to an unexpected value for either
masterStats.subtaskStartCount or masterStats.subtaskEndCount or for an
unexpected value for either subtaskStats.chunksOut or
subtaskStats.chunksIn. I suspect that the underlying issue is the
expectation of the tests with respect to when a sink is retired by the
master and whether or not a new subtaskStats object is allocated or an
old one reused (e.g., because the subtask was not closed by an idle
timeout or the like).
     
These test failures can be cleared up if you uncomment the [tmp] list
and its use in {@link
AbstractMasterStats#getSubtaskStats(Object)}. This forces the internal
map to hold a hard reference to the subtask statistics objects. This
demonstrates that the problem is not with the termination conditions
for the {@link AbstractMaster}. However, the unit test are still
broken until I track down the underlying assumption within them which
is being violated.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1774 8f7bc0f5-282e-0410-95e3-8d296e9bb460