Added statistics on the distributed join evaluation to the rule execution log, and added additional counters for the #of index partitions spanned by the join dimension, for the access path used for that join dimension, and a timestamp at the start of the join detail record for help in correlating the join statistics to the performance counters.
Added an alternative asynchronous statement buffer implementation which should resolve a bottleneck in the distributed data load on the clients.  The symptom was that all parsers would be blocked after a while, causing progress to depend on TERM2ID chunk timeouts.  This feature has not been tested yet.  It can be disabled by using the previous version of the RDFFileLoadTask class (a minor edit).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1828 8f7bc0f5-282e-0410-95e3-8d296e9bb460