added notes on unix line endings for scripts.
modified the lubm integration class to define three run modes (Generate, GenerateAndLoad, and Load).  This is being done since the lubm generator can not keep up with the parsers and the database (it is single threaded).

updated configuration files to reflect the lubm generator change.

increased heap on the CS for the 3-node cluster to 1600m.

added the average active count for the parser thread pool to the port-mortem counter set queries.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@1857 8f7bc0f5-282e-0410-95e3-8d296e9bb460