The mapped task master is now running in a standalone federation (single machine).  I am committing changes to test this out on a cluster.
Modified the JavaServicesConfiguration to specify the JAVA_HOME property rather than the path of the java executable.

Modified the service starter classes to generate a startService file in the service directory which contains a script that can be used to start the service in question.  This should make it easier to debug cluster configurations.

Added a "standalone federation" configuration.  This is a full distributed jini-based federation, but all services run on the local host.  This is intended for debugging in an RMI environment.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@1986 8f7bc0f5-282e-0410-95e3-8d296e9bb460