Lot's more work on the B+Tree refactor.  There is still a remaining bug with search on the front-coded raba (the boundary condition is where the probe key lies before the first bucket).  The rest of the raba coders and the node/leaf coders are fine (the node/leaf coders still need to be optimized for their footprint on the disk).
The next step is to reconcile with the historical use of the IDataSerializer.  This will touch NodeSerializer (which is currently broken so this branch does not compile), ITupleSerializer, the IIndexProcedure implementations, etc.  This will also provide the use of the coded keys/values representations for chunked iterator result sets and index procedures.  At that point we can start end-to-end testing.

Note: The FastRDFKeyValueSerializer may be broken right now.  I will test that once the NodeSerializer is fixed.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BTREE_BUFFER_BRANCH@2032 8f7bc0f5-282e-0410-95e3-8d296e9bb460