Fixed bug where a Temporary BufferMode, which is backed by disk, would not be extended beyond 2G (Integer#MAX_VALUE).
Also fixed a bug where a request larger than the initial extent could result in an extension of the store smaller than the required amount.

Added BufferMode#isFullyBuffered() to exactly capture the semantics of a store with an image of its contents in memory, which is what motivates the upper bound of 2G on a Transient store (but not on a Temporary store).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BTREE_BUFFER_BRANCH@2046 8f7bc0f5-282e-0410-95e3-8d296e9bb460