This commit fixes problems with failing to update the min/max tuple revision timestamp on a leaf, but does not propagate the min/max tuple revisions to the node yet.  The node data record will actually have to store per-child min/max tuple revision timestamps to support fast recalculation of the min/max for the node when a child is deleted (otherwise we have to scan the children).  Therefore I have added a new configuration option to enable tuple revision filtering, and the node will only track revision timestamp information if that flag is set.
The fix to the Leaf to maintain the min/max tuple revision timestamp also fixed TestIsolatedFusedView and TestTx#test_writeIsolation(). The problem was that when min==max we do not record the per-tuple timestamps since they must be zero.

Fixed a problem with AsynchronousOverflowTask where it would shutdown the federation's thread pool if it ran its subtasks with full concurrency.

Fixed a problem where a Temporary (vs Transient) store was not extended beyond 2GB.

There is still a problem with TestTransientBTree#delete().

TestRangeQuery#reverseScan() and #removeAll() were fixed by some other changes.

Fixed NPE in AbstractBTreeTupleCursor - it was failing to test fromKey!=null and toKey!=null when range checking a key.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2047 8f7bc0f5-282e-0410-95e3-8d296e9bb460