Commented out "fail()" for some NOP unit tests, leaving only @todo to indicate that they need to be written.
Modified the EmptyRabaCoder to persist the size of the raba.

There are only two LTS tests which are failing, both of which are in TestTMUtility.

junit.framework.AssertionFailedError
	at junit.framework.Assert.fail(Assert.java:47)
	at junit.framework.Assert.assertTrue(Assert.java:20)
	at junit.framework.Assert.assertTrue(Assert.java:27)
	at com.bigdata.rdf.rules.TestTMUtility.test_rewrite1(TestTMUtility.java:91)

junit.framework.ComparisonFailure: expected:<focusStor...> but was:<databas...>
	at junit.framework.Assert.assertEquals(Assert.java:81)
	at junit.framework.Assert.assertEquals(Assert.java:87)
	at com.bigdata.rdf.rules.TestTMUtility.test_rewrite2(TestTMUtility.java:128)

These unit tests were broken by revision# 1980, which included some magic predicate related edits to TMUtility.
	

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BTREE_BUFFER_BRANCH@2051 8f7bc0f5-282e-0410-95e3-8d296e9bb460