done. Modified AbstractBTree, NodeSerializer, IRabaCoder, and IAbstractNodeCoder to support an "encodeLive()" method which returns the coded object wrapping its data record.  For most IRabaCoders, this is significantly faster than decoding to obtain a live ICodedRaba.
done. Added ByteArrayBuffer#trim(), which replaces the backing buffer
with an exact fit byte[] and returns the old backing buffer.  This was
used to optimize the common code path in AbstractBTree and
NodeSerializer where we immediately require access to the coded node
or leaf and the data was coded onto a shared buffer.

done. The CanonicalHuffmanRabaCoder's CodedRaba no longer directly
addresses into a held reference to the backing byte[] because the
backing byte[] reference gets replaced by ByteArrayBuffer#trim().

done. Fixed bug in the FastRDFValueCoder (was FastRDFValueCompressor)
where it was unable to use random access. The problem was invoking
InputBitStream#skip() rather than #position().  The former is only for
sequential access.

done. Modified AbstractBTreeTestCase#assertSameRaba() to test random
access using a randomized ordering.

done. The CanonicalHuffmanRabaCoder's iterator() could copy the data
into a backing ByteArrayBuffer and then use toByteArray() to return a
clone of the data. This would be faster since it would avoid
double-scans of the coded symbols.

done. getByte() and symbol2byte() are 12% of the canonical huffman
raba coder total time!  get an array reference on entry to any API
level method and pass through the stack to do direct array operations
whenever possible.

done. Some errors were showing up on U10 and U5 which were not present
in the test suite.  These were tracked down to the FastRDFValueCoder.
It was failing with the modification to use ByteArrayBuffer#trim(),
though I am not clear why.  I modified the FastRDFValueCoder to write
out the version and the size of the raba without using the OBS in
order to simplify the logic.  It is now running correctly on both the
test cases and on those data sets.

done. Fixed StatementEnum#isOverride().  The method was testing "(b &
StatementEnum.MASK_OVERRIDE) == 1", which is never true.  It should
have been testing "(b & StatementEnum.MASK_OVERRIDE) != 0".

TODO: There is a TM bug. The KB will not downgrade an Explicit statement to Inferred.  There is now a unit test for this in TestTruthMaintenance.  I am looking into why this is failing.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2054 8f7bc0f5-282e-0410-95e3-8d296e9bb460