Renamed InferenceEngine.Options.FORWARD_CHAIN_OWL_TRANSITIVE_PROPERY to correct the spelling error (PROPER_T_Y).
Move the initialization of the "IN" constraint for the default and named graph queries out of the getAcessPath() method on the NamedGraphSolutionExpander and DefaultGraphSolutionExpander.  This was paying a huge performance penalty when the access path was being requested a large number of times during JOIN subquery evaluation.

Refactored the DistinctSPOIterator into its own top-level class.

Modified the DistinctSPOIterator to use a hash set unless the set size
grows to be very large.  Further modified to use "view" which spills
over only the new inserts into the btree.  The hash set is much faster
and lighter than the BTree.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BTREE_BUFFER_BRANCH@2126 8f7bc0f5-282e-0410-95e3-8d296e9bb460