bug fix to handles queries such as
	SELECT * WHERE {graph ?g {?g :p :o } }

These queries were being run against a default graph defined as the RDF Merge of all graphs in the quad store.  As such, the context was being stripped from each quad and the access path was filtering for the distinct (s,p,o) occurrences.  The code path was modified to run against all graphs in the store instead -- just as if you had explicitly listed FROM NAMED for each graph in the store.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BTREE_BUFFER_BRANCH@2128 8f7bc0f5-282e-0410-95e3-8d296e9bb460