Bug fix imposes a constraint on an access path if a variable is reused within a predicate.  For example, in (?g, p1, o1, ?g), the variable "g" is used twice.  Added the SameVariableConstraint class to support this and integrated this into the AbstractAccessPath.  The filter will be evaluated at the data, so rejected elements will not be materialized across an RMI boundary.
Bug fix for Constant.equals(Object).  The method was not implemented
with Object and equals() was therefore not being overridden when
comparing two constants, a constant and a variable, etc.

Added IPredicate#get(E e, int index).  This method allows you to
access the value of the element corresponding to the specified index
in the predicate.

Deprecated IPredicate#isFullyBound(), which does not handle relations
having indices whose keys are formed from only a subset of the slots
of the predicate.

Added IPredicate#isFullyBound(IKeyOrder) and
IPredicate#getVariableCount(IKeyOrder).  These method are aware of
which slots in the predicate are required in order to form the key for
the specified key order. 

Added IKeyOrder#getKeyArity(), IKeyOrder#getKeyOrder(int keyPos).
These methods make it possible to write
IPredicate#isFullyBound(IKeyOrder) and
IPredicate#getVariableCount(IKeyOrder) implementations.

Modified SPOPredicate#isFullyBound() and
SPOPreciate#getVariableCount() to include [c] in their consideration
iff c != null, but removed all use of those methods from the code
base.

LexiconRelation#getKeyOrder() -- this lacks the concept of the key
positions in the BigdataValue.  They are the RDF Value and the term
identifier.  Assigned 0 to the value and 1 to the term identifier.

Factored the "name" of the LexiconKeyOrder out of its serializable
state.

Bug fix to RDF test suite, which was not running the LocalTripleStore in its provenance mode.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2140 8f7bc0f5-282e-0410-95e3-8d296e9bb460