Bug fix for NamedLock where it was not detecting a lock timeout (failure to check the return code for tryLock()).  Also modified the interface to throw the TimeoutException.  This change also effected the IResourceLockService.  The ZK impl also was ignoring the timeout parameter.
various javadoc edits.

Added stress test from MattR which fails on centos 5 64-bit with JDK 1.6.0_16.  No diagnosis for the failure yet.  I suspect a problem with the locator cache, perhaps in the ConcurrentHashMap on the platform.

The AbstractPendingSetMasterTask was not adding anything to its pending set (the method was not being called).  

Modified the ResourceBufferTask to always use a hash map.  There a problems when the resource is a File or URL which need to be resolved by consistent coding of instances of that class in the BigdataMap keys.  Flagged this issue in the javadoc.  It throws an UnsupportedOperationException if you request the BigdataMap instance for the mapped job and now defaults to using a standard LinkedHashMap.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2171 8f7bc0f5-282e-0410-95e3-8d296e9bb460