Added Options to the TemporaryStoreFactory to control the location of the temporary stores, their offset bits, and their maximum extent before a new instance is created.
Added/modified constructors to the TemporaryStoreFactory to use System.getProperty() for the zero argument constructor, the caller's Properties, or the caller's specific values for the temporary directory, offset bits, and maximum extent.

Modified the IBigdataClient.Options, AbstractClient, and AbstractFederation to use the same configuration options and defaults for the TemporaryStoreFactory attached to the client/service.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BTREE_BUFFER_BRANCH@2203 8f7bc0f5-282e-0410-95e3-8d296e9bb460