Various bug fixed identified by find bugs.
Modified the IndexSegmentBuilder to optionally stuff the generated node and leaf data records for the new index segment into the shared LRU as it goes.  This option can be disabled using a configuration property defined by the LRUNexus.

Pulled the getFromKey() and getToKey() methods out of SPOAccessPath - they are now part of SPOKeyOrder.

The IGlobalLRU interface was modified to operate in terms of the UUID of the IRawStore and the IAddressManager interface.  The other facets of the IRawStore (it's implementation class, the backing file, etc.) are no longer available to the IGlobalLRU implementations.  This was done so the IndexSegmentBuilder could create a correct ILRUCache instance for the new IndexSegmentStore before the store exists.  A static method on LRUNexus now encapsulates the logic to obtain an ILRUCache when you have an IRawStore instance.

Fixed a formatting problem in RuleStats which was polluting the rule execution log.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2230 8f7bc0f5-282e-0410-95e3-8d296e9bb460