Changed the partitionSizeMultipler back to 1.  This results in shards of approximately 200MB, which is the desired shard size.
Changed the #of clients from 4 to 3 and increased the number of data services.

Disabled the forceOverflow option for the RDF Data load.  Forcing overflow was triggering too many concurrent build/merge tasks, some of which did not complete in a timely manner.

All of these changes help to avoid the journal extent getting out of control and result in more stable sustained write performance.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2295 8f7bc0f5-282e-0410-95e3-8d296e9bb460