Fix to the BTree to increase the potential concurrency of AbstractBTree#touch(AbstractNode).  I've run the test suites against this but I have not yet run performance benchmarks designed to quantify or tune the new solution.  In brief, for a read-only B+Tree the writeRetentionQueue uses thread-local queues to batch updates to the shared (inner) queue.
This involved a variety of changes, including raising the readOnly parameter into the BTree constructor signature.  One consequence of this set of changes is that the BTree from the last commit time is no longer placed into the set of historical btrees for the new journal by closeForWrites().  Getting this to work safely will take a bit more work, but note that if the global LRU is enabled, then the data records will all be buffered anyway.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/DEV_BRANCH_27_OCT_2009@2306 8f7bc0f5-282e-0410-95e3-8d296e9bb460