Added more information to the DataLoader to report on the nodes, leaves and bytes written on each index when INFO logging is enabled for that class.
Made the default value of the B+Tree write retention queue capacity a function of the maximum Java heap size (-Xmx) for better ergonomics.

Added javadoc comment to DefaultEvaluationPlan2 indicating that the recent change in the join planner actually hurts performance on LUBM Q2.  We are looking into that now.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2354 8f7bc0f5-282e-0410-95e3-8d296e9bb460