Modified AbstractTripleStore#newJoinNexusFactory(...) to accept a Properties object containing overrides (aka query hints).
Tracked down a bug with MattR in the REVERSE tuple cursor.  The error arises when the toKey is not found.  The tuple cursor starts from one index too high in the leaf and rangeCheck() then fails the current tuple, causing it to skip the leaf.  This has not been fixed yet, but the unit tests have been modified and now fail the REVERSE tuple cursor for this set of conditions.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/DEV_BRANCH_27_OCT_2009@2373 8f7bc0f5-282e-0410-95e3-8d296e9bb460