Bug fix for REVERSE cursor traversal.  There were two separate conditions which were not being handled correctly.
(1) When the toKey did not exist, an insertion point was returned by keys.search().  The code was failing to increment the insertion point with the result that the insertion point (when turned into a within the tuple index) was one beyond the desired starting point.  rangeCheck() would then fail the current tuple.  This is now handled in lastPosition().

(2) When the toKey did not exist and it was the first key in the leaf, then the lastPosition() method should have advanced the cursor to the last tuple in the prior leaf (if any).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2382 8f7bc0f5-282e-0410-95e3-8d296e9bb460