Added a low-level feature to report out ISPO#isModified() for SPORelation#insert(...) and SPORelation#delete(...).
Note: This flag does not make it through the StatementBuffer due to the manner in which that class is structured.  Basically, the original Statements flowing into the StatementBuffer become disassociated with the SPOs used to write on the database. This goes back to its history as a parser driven API.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/DEV_BRANCH_27_OCT_2009@2388 8f7bc0f5-282e-0410-95e3-8d296e9bb460