Removed IUpdateStore.
Modified the IndexSegmentBuilder to directly write leaves onto the output file (using a write cache to speed this up) and to optionally buffer nodes in RAM until they are written out (again, using a write cache).

Corrected the write cache API to indicate that you read and write with a file offset, not a record address.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2442 8f7bc0f5-282e-0410-95e3-8d296e9bb460