Ok.  I have tracked down the problem with the scale-out joins.  Whenever a relation is resolved, it is resolved used the federation's index manager since that has access to the global row store, which is where we store the relation's metadata.  This means that (for scale-out) relation#getIndexManager() ALWAYS returns the federation's index manager.
However, the scale-out pipeline join depends on the data service's local index manager in order to access the specific shard on which it needs to read for a given join dimension.

This problem was introduced a few weeks ago when I was doing some performance tuning for BSBM.  The solution is just to pass the RDFJoinNexus's indexManager (which is the data service local index manager) in, e.g., using SPORelation#getAccessPathForIndexPartition(indexManager, predicate).  The logic to instantiate the access path for the relation type and the appropriate index manager are now co-located again.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches@2456 8f7bc0f5-282e-0410-95e3-8d296e9bb460